requirements.txt
PyYAML~=6.0
pandas~=2.0.1
tqdm~=4.65.0
numpy~=1.24.3
setuptools~=60.2.0